moved test to semantically correct place
